[FIX] : CURRENT_PID 값이 있을 경우에만 Terminate(정상 종료 `-15`) 처리하도록 변경

+++ scripts/switch.sh
@@ -24,4 +24,6 @@ echo "> Nginx reloaded."
 
 CURRENT_PID=$(lsof -Fp -i TCP:${CURRENT_PORT} | grep -Po 'p[0-9]+' | grep -Po '[0-9]+')
 
-sudo kill -15 ${CURRENT_PID}
+if [[ -n ${CURRENT_PID} ]]; then
+  sudo kill -15 ${CURRENT_PID}
+fi
